feat: a working basic authentication

+++ server/src/api/services/projects.service.ts
@@ -1,13 +1,16 @@
 import { ProjectDocument, ProjectModel as Project } from '../models/project.model';
 
-const selectString = '_id title url type children issues';
+const selectString = '_id title url type children createdBy issues';
 
-export const getAllProjectsService = async () => {
-  const query = await Project.find().select(selectString).exec();
+// type ProjDocument = Omit<ProjectDocument, "createdBy">
+// type MyUserDocument = Pick<ProjectDocument, "createdBy">
+
+export const getAllProjectsService = async (userId: string) => {
+  const query = await Project.find({ createdBy: userId}).select(selectString).exec();
   return query;
 }
 
-export const createOneProjectService = async (requestBody: ProjectDocument): Promise<ProjectDocument> => {
+export const createOneProjectService = async (requestBody: ProjectDocument, userId: string): Promise<ProjectDocument> => {
   const project = new Project({
     title: requestBody.title,
     url: requestBody.url,
@@ -16,17 +19,18 @@ export const createOneProjectService = async (requestBody: ProjectDocument): Pro
     issues: {
       url: requestBody.issues.url,
     },
+    createdBy: userId,
   });
   const save = await project.save();
   return save;
 }
 
-export const getOneProjectService = async (paramsId: string) => {
-  const query = Project.findById(paramsId).select(selectString).exec();
+export const getOneProjectService = async (paramsId: string, userId: string) => {
+  const query = Project.findOne({_id: paramsId, createdBy: userId}).select(selectString).exec();
   return query;
 }
 
-export const deleteOneProjectService = async (paramsId: string) => {
-  const query = await Project.findOneAndDelete({ _id: paramsId }).exec();
+export const deleteOneProjectService = async (paramsId: string, userId: string) => {
+  const query = await Project.findOneAndDelete({_id: paramsId, createdBy: userId}).exec();
   return query;
 }
